Refactor data types from double to float for optimization; update button, LED, pot, and sequence player

+++ include/hardware/led.h
@@ -7,16 +7,16 @@ class LED
 {
 private:
     int pin;
-    double blinkStart;    // Time accumulated since blink started in seconds
-    double blinkDuration; // Duration of blink in seconds
+    float blinkStart;    // Time accumulated since blink started in seconds
+    float blinkDuration; // Duration of blink in seconds
     bool isBlinking;
 
 public:
     LED(int ledPin);
     void on();
     void off();
-    void blink(double duration);
-    void update(double dt); // dt is delta time in seconds
+    void blink(float duration);
+    void update(float dt); // dt is delta time in seconds
 };
 
 #endif // LED_H
